A number of code simplifications (removing redundant casts, unneeded toStrings, redundant checks, simplifying a number of string operations, consolidated duplicated switch cases. The main non-simplification is adding final where possible, which allows for better optimizations (in the compiler)

+++ WEB-INF/classes/dods/dap/Int16PrimitiveVector.java
@@ -179,7 +179,7 @@ public class Int16PrimitiveVector extends PrimitiveVector implements Cloneable {
   @Override
   public void externalize(DataOutputStream sink) throws IOException {
     for (short val : vals) {
-      sink.writeInt((int) val);
+      sink.writeInt(val);
     }
   }
 
@@ -194,7 +194,7 @@ public class Int16PrimitiveVector extends PrimitiveVector implements Cloneable {
    */
   public void externalize(DataOutputStream sink, int start, int stop, int stride)
       throws IOException {
-    for (int i = start; i <= stop; i += stride) sink.writeInt((int) vals[i]);
+    for (int i = start; i <= stop; i += stride) sink.writeInt(vals[i]);
   }
 
   /**
